third_party/WebKit/LayoutTests/platform/android/http/tests/media/media-source/mediasource-config-change-mp4-av-audio-bitrate-expected.txt

CONSOLE WARNING: line 95: Ignore this warning. See https://crbug.com/568704#c2
This is a testharness.js-based test.
PASS Tests mp4 audio bitrate changes in multiplexed content. 
Harness: the test ran to completion.
